pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mesg-agent"
version = "1.0.0"
description = "AI Agent with Feishu integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "agent", "feishu", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "lark-oapi>=1.5.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/mesg-agent"
Repository = "https://github.com/yourusername/mesg-agent"
Issues = "https://github.com/yourusername/mesg-agent/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::UserWarning:pkg_resources.*",
    "ignore::DeprecationWarning:pkg_resources.*",
    "ignore::DeprecationWarning:lark_oapi.*",
    "ignore::DeprecationWarning:websockets.*",
    "ignore::UserWarning:lark_oapi.*",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["core"]
